requirements.txt
alembic
asyncpg
psycopg
psycopg-binary
sqlalchemy

fastapi
pydantic
pydantic-settings
uvicorn

natasha
setuptools
pymorphy2
python-docx
pymupdf

requests